Move WorkflowState model to workflows submodule
